Installed `prettier` with `husky` & `lint-staged` to format our code on commits. Formatted all our current code with `prettier`:
    $ npx prettier --config ./.prettierrc.json --write *.js gfp/*/**.js

Needed to format `gfp/__tests__/test_utils.js` manually so the indents are correct. Fixed other linter errors.

Some of these are in `Adblock Plus`'s old code. Used to want to touch this code as little as possible. But doesn't matter anymore, as not like we'll update this to the latest `Adblock Plus` anyway. We don't even use `Adblock Plus` anymore in favor of `uBlock Origin`.